services/api-service/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
requests==2.32.3
psycopg2-binary==2.9.10
